+++ snack3_demo/src/test/java/benchmark/BaseSpeedTest2.java
@@ -1,4 +1,4 @@
-package speed;
+package benchmark;
 
 import org.junit.Test;
 import org.noear.snack.ONode;
