+++ snack3_demo/src/test/java/benchmark/GsonDateSerialize.java
@@ -1,4 +1,4 @@
-package speed;
+package benchmark;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonPrimitive;
