+++ snack3_demo/src/test/java/benchmark/SpeedFastjsonTest.java
@@ -1,4 +1,4 @@
-package speed;
+package benchmark;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.SerializerFeature;
