+++ snack3_demo/src/test/java/benchmark/jsonpath/SpeedFastjsonJsonPathTest.java
@@ -1,29 +1,26 @@
-package speed.jsonpath;
+package benchmark.jsonpath;
 
-import com.jayway.jsonpath.Configuration;
-import com.jayway.jsonpath.JsonPath;
-import com.jayway.jsonpath.ReadContext;
-import net.minidev.json.JSONArray;
+import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONArray;
+import com.alibaba.fastjson.JSONObject;
+import com.alibaba.fastjson.JSONPath;
 import org.junit.Test;
 
 
-public class SpeedJaywayJsonPathTest {
+public class SpeedFastjsonJsonPathTest {
 
 
     @Test
     public void test1(){
-        //1000000=>2658,2633,2590
+        //1000000=>529,546,539
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray rst = context.read("$..a");
-        assert rst.size() ==2;
+        JSONObject obj = JSON.parseObject(text);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$..a");
+            JSONPath.eval(obj,"$..a");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -35,17 +32,17 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test2(){
-        //1000000=>3227,3220,3156
+        //不支持
+        //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
+        JSONObject obj = JSON.parseObject(text);
 
-        JSONArray rst = context.read("$..*");
-        assert rst.size() == 16;
+        Object tmp  =JSONPath.eval(obj,"$..*");;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$..*");
+            JSONPath.eval(obj,"$..*");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -57,17 +54,17 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test3(){
-        //1000000=>782,798,776
+        //1000000=>85,90,86
+        //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
+        JSONObject obj = JSON.parseObject(text);
 
-        JSONArray rst = context.read("$.data.list[1,4]");
-        assert rst.size() == 2;
+        Object tmp = JSONPath.eval(obj,"$.data.list[1,4]");
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$.data.list[1,4]");
+            JSONPath.eval(obj,"$.data.list[1,4]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -79,18 +76,15 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test4(){
-        //1000000=>941,899,947
+        //1000000=>105,109,114
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray rst = context.read("$.data.list[1:4]");
-        assert rst.size() == 3;
+        JSONObject obj = JSON.parseObject(text);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$.data.list[1:4]");
+            JSONPath.eval(obj,"$.data.list[1:4]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -102,19 +96,16 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test41(){
-        //1000000=>2522,2551,2591
+        //1000000=>325,321,319
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray rst = context.read("$..ary2[0].a");
-
-        assert rst.size() == 1;
+        JSONObject obj = JSON.parseObject(text);
+        Object tmp = JSONPath.eval(obj,"$..ary2[0].a");
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$..ary2[0].a");
+            JSONPath.eval(obj,"$..ary2[0].a");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -125,30 +116,16 @@ public class SpeedJaywayJsonPathTest {
     }
 
     @Test
-    public void test42() {
-        //1000000=>5494,5326,5483
+    public void test42(){
+        //出错
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
-
-//        Object tmp00 = JsonPath.read(obj,"$..ary2");
-//        Object tmp01 = JsonPath.read(obj,"$..list[2]");
-//        Object tmp0 = JsonPath.read(obj,"$..list[-2:]");
-//        Object tmp1 = JsonPath.read(obj,"$..ary2[0].a");
-//
-//        Object tmp2 = JsonPath.read(obj,"$.data.list[?(@ == $..ary2[0].a.min())]");
-
-        JSONArray tmp30 = context.read("$..ary2[0].a");
-        assert tmp30.size() == 1;
-
-        JSONArray tmp3 = context.read("$.data.list[?(@ in $..ary2[0].a)]");
-        assert tmp3.size() == 1;
-
+        JSONObject obj = JSON.parseObject(text);
 
         long start = System.currentTimeMillis();
-        for (int i = 0, len = 1000000; i < len; i++) {
-            context.read("$.data.list[?(@ == $..ary2[0].a)]");
+        for(int i=0,len=1000000; i<len; i++) {
+            JSONPath.eval(obj,"$.data.list[?(@ in $..ary2[0].a)]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -160,18 +137,17 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test5(){
-        //1000000=>929,826,837
+        //1000000=>60,58,58
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
+        JSONObject obj = JSON.parseObject(text);
 
-        Object tmp = context.read("data.ary2[1].b.c");;
-        assert "ddd".equals(tmp);
+        Object tmp =JSONPath.eval(obj,"$.data.ary2[1].b.c");
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$.data.ary2[1].b.c");
+            JSONPath.eval(obj,"$.data.ary2[1].b.c");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -183,18 +159,15 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test6(){
-        //1000000=>1105,1025,1050
+        //不支持*
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray tmp = context.read("$.data.ary2[*].b.c");
-        assert tmp.size() == 1;
+        JSONObject obj = JSON.parseObject(text);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$.data.ary2[*].b.c");
+            JSONPath.eval(obj,"$.data.ary2[*].b.c"); //不支持*
         }
 
         long times = System.currentTimeMillis() - start;
@@ -206,18 +179,15 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test7(){
-        //1000000=>5628,5739,5636
+        //不支持
         //
         //1.加载json
         String text = ("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray tmp = context.read("$..b[?(@.c == 12)]");
-        assert tmp.size() == 0;
+        JSONArray obj = JSONArray.parseArray(text);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$..b[?(@.c == 12)]");
+            JSONPath.eval(obj,"$..b[?(@.c == 12)]");//不支持
         }
 
         long times = System.currentTimeMillis() - start;
@@ -229,17 +199,16 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test8(){
-        //运行会出错
+        //不支持
+        //
         //1.加载json
-        String text =("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
-        ReadContext context = JsonPath.parse(text);
+        String text = ("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
+        JSONArray obj = JSONArray.parseArray(text);
 
-        JSONArray tmp = context.read("$..c");
-        assert tmp.size() == 3;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$..c.min()");
+            JSONPath.eval(obj,"$..c.min()");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -251,18 +220,13 @@ public class SpeedJaywayJsonPathTest {
 
     @Test
     public void test9(){
-        //1000000=>3575,3591,3813
-        //
         //1.加载json
         String text = ("[{c:'aaaa'}, {b:'cccc'}, {c:'cccaa'}]");
-        ReadContext context = JsonPath.parse(text);
-
-        JSONArray tmp = context.read("$[?(@.c =~ /.*a+/)]");
-        assert tmp.size() == 2;
+        JSONArray obj = JSONArray.parseArray(text);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            context.read("$[?(@.c =~ /a+/)]");//
+            JSONPath.eval(obj,"$[?(@.c =~ /a+/)]");//不支持
         }
 
         long times = System.currentTimeMillis() - start;
