+++ snack3_demo/src/test/java/benchmark/jsonpath/SpeedJaywayJsonPathTest.java
@@ -1,26 +1,28 @@
-package speed.jsonpath;
+package benchmark.jsonpath;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONArray;
-import com.alibaba.fastjson.JSONObject;
-import com.alibaba.fastjson.JSONPath;
+import com.jayway.jsonpath.JsonPath;
+import com.jayway.jsonpath.ReadContext;
+import net.minidev.json.JSONArray;
 import org.junit.Test;
 
 
-public class SpeedFastjsonJsonPathTest {
+public class SpeedJaywayJsonPathTest {
 
 
     @Test
     public void test1(){
-        //1000000=>529,546,539
+        //1000000=>2658,2633,2590
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray rst = context.read("$..a");
+        assert rst.size() ==2;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$..a");
+            context.read("$..a");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -32,17 +34,17 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test2(){
-        //不支持
-        //
+        //1000000=>3227,3220,3156
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
 
-        Object tmp  =JSONPath.eval(obj,"$..*");;
+        JSONArray rst = context.read("$..*");
+        assert rst.size() == 16;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$..*");
+            context.read("$..*");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -54,17 +56,17 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test3(){
-        //1000000=>85,90,86
-        //
+        //1000000=>782,798,776
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
 
-        Object tmp = JSONPath.eval(obj,"$.data.list[1,4]");
+        JSONArray rst = context.read("$.data.list[1,4]");
+        assert rst.size() == 2;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$.data.list[1,4]");
+            context.read("$.data.list[1,4]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -76,15 +78,18 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test4(){
-        //1000000=>105,109,114
+        //1000000=>941,899,947
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray rst = context.read("$.data.list[1:4]");
+        assert rst.size() == 3;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$.data.list[1:4]");
+            context.read("$.data.list[1:4]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -96,16 +101,19 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test41(){
-        //1000000=>325,321,319
+        //1000000=>2522,2551,2591
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
-        Object tmp = JSONPath.eval(obj,"$..ary2[0].a");
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray rst = context.read("$..ary2[0].a");
+
+        assert rst.size() == 1;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$..ary2[0].a");
+            context.read("$..ary2[0].a");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -116,16 +124,30 @@ public class SpeedFastjsonJsonPathTest {
     }
 
     @Test
-    public void test42(){
-        //出错
+    public void test42() {
+        //1000000=>5494,5326,5483
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
+
+//        Object tmp00 = JsonPath.read(obj,"$..ary2");
+//        Object tmp01 = JsonPath.read(obj,"$..list[2]");
+//        Object tmp0 = JsonPath.read(obj,"$..list[-2:]");
+//        Object tmp1 = JsonPath.read(obj,"$..ary2[0].a");
+//
+//        Object tmp2 = JsonPath.read(obj,"$.data.list[?(@ == $..ary2[0].a.min())]");
+
+        JSONArray tmp30 = context.read("$..ary2[0].a");
+        assert tmp30.size() == 1;
+
+        JSONArray tmp3 = context.read("$.data.list[?(@ in $..ary2[0].a)]");
+        assert tmp3.size() == 1;
+
 
         long start = System.currentTimeMillis();
-        for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$.data.list[?(@ in $..ary2[0].a)]");
+        for (int i = 0, len = 1000000; i < len; i++) {
+            context.read("$.data.list[?(@ == $..ary2[0].a)]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -137,17 +159,18 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test5(){
-        //1000000=>60,58,58
+        //1000000=>929,826,837
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
 
-        Object tmp =JSONPath.eval(obj,"$.data.ary2[1].b.c");
+        Object tmp = context.read("data.ary2[1].b.c");;
+        assert "ddd".equals(tmp);
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$.data.ary2[1].b.c");
+            context.read("$.data.ary2[1].b.c");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -159,15 +182,18 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test6(){
-        //不支持*
+        //1000000=>1105,1025,1050
         //
         //1.加载json
         String text = ("{code:1,msg:'Hello world',data:{list:[1,2,3,4,5], ary2:[{a:2},{a:3,b:{c:'ddd'}}]}}");
-        JSONObject obj = JSON.parseObject(text);
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray tmp = context.read("$.data.ary2[*].b.c");
+        assert tmp.size() == 1;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$.data.ary2[*].b.c"); //不支持*
+            context.read("$.data.ary2[*].b.c");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -179,15 +205,18 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test7(){
-        //不支持
+        //1000000=>5628,5739,5636
         //
         //1.加载json
         String text = ("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
-        JSONArray obj = JSONArray.parseArray(text);
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray tmp = context.read("$..b[?(@.c == 12)]");
+        assert tmp.size() == 0;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$..b[?(@.c == 12)]");//不支持
+            context.read("$..b[?(@.c == 12)]");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -199,16 +228,17 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test8(){
-        //不支持
-        //
+        //运行会出错
         //1.加载json
-        String text = ("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
-        JSONArray obj = JSONArray.parseArray(text);
+        String text =("[{b:{c:1}}, {b:{d:1}}, {b:{c:2}}, {b:{c:23}}]");
+        ReadContext context = JsonPath.parse(text);
 
+        JSONArray tmp = context.read("$..c");
+        assert tmp.size() == 3;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$..c.min()");
+            context.read("$..c.min()");
         }
 
         long times = System.currentTimeMillis() - start;
@@ -220,13 +250,18 @@ public class SpeedFastjsonJsonPathTest {
 
     @Test
     public void test9(){
+        //1000000=>3575,3591,3813
+        //
         //1.加载json
         String text = ("[{c:'aaaa'}, {b:'cccc'}, {c:'cccaa'}]");
-        JSONArray obj = JSONArray.parseArray(text);
+        ReadContext context = JsonPath.parse(text);
+
+        JSONArray tmp = context.read("$[?(@.c =~ /.*a+/)]");
+        assert tmp.size() == 2;
 
         long start = System.currentTimeMillis();
         for(int i=0,len=1000000; i<len; i++) {
-            JSONPath.eval(obj,"$[?(@.c =~ /a+/)]");//不支持
+            context.read("$[?(@.c =~ /a+/)]");//
         }
 
         long times = System.currentTimeMillis() - start;
