+++ snack3_demo/src/test/java/benchmark/jsonpath/SpeedJsonPathTest2.java
@@ -1,4 +1,4 @@
-package speed.jsonpath;
+package benchmark.jsonpath;
 
 import org.junit.Test;
 import org.noear.snack.ONode;
